requirements.txt
factpages-py
streamlit
anthropic
pandas
plotly
python-dotenv
requests
